tried a different logic with delete_repository

+++ ui_objects/project.py
@@ -54,8 +54,8 @@ class Project(QWidget):
         self.__create_repo_button.clicked.connect(self.__create_repository_popup)
 
         # Delete button for the repository
-        self.__delete_button = QPushButton("X")
-        self.__delete_button.clicked.connect(self.__delete_repository)
+        self._delete_button = QPushButton("X")
+        self._delete_button.clicked.connect(self.delete_repository)
 
         # adding the button and the repos layout to the main project layout
         self.__layout.addWidget(self.__create_repo_button)
@@ -103,9 +103,11 @@ class Project(QWidget):
         :param num:
         :return:
         """
-        repo.deleteLater()
-        self.repository_list.remove(repo)
-        logging.debug(f"Deleted repository {repo} / Remaining repositories: {len(self.repository_list)}")
+        if 0 <= num < len(self.__repository_list):
+            repo = self.__repository_list[num]
+            repo.deleteLater()
+            del self.__repository_list[num]
+            logging.debug(f"Deleted repository {repo} / Remaining repositories: {len(self.__repository_list)}")
 
     def to_dict(self):
         """
